feat: automata
