+++ automata.py
@@ -1,53 +1,26 @@
+import players
 from abc import ABC, abstractmethod
 
 
-class Player(ABC):
-    def __init__(self, nickname):
-        self.nickname = nickname
+class State:
+    def __init__(self, game):
+        self.game = game
 
     @abstractmethod
-    def talk_time(self):
+    def handle(self):
+        pass
+
+    @abstractmethod
+    def set_state(self, state):
         pass
 
 
-class Citizen(Player):
-    def __init__(self, nickname):
-        super(Citizen, self).__init__(nickname)
+class CheckRoles(State):
+    def __init__(self, state):
+        super(CheckRoles, self).__init__(state)
 
-    def talk_time(self):
+    def handle(self):
         pass
 
-
-class Sheriff(Citizen):
-    def __init__(self, nickname):
-        super(Sheriff, self).__init__(nickname)
-
-    def mafia_check(self, nickname):
+    def set_state(self, state):
         pass
-
-
-class Mafia(Player):
-    def __init__(self, nickname):
-        super(Mafia, self).__init__(nickname)
-
-    def talk_time(self):
-        pass
-
-    def kill_player(self, nickname):
-        pass
-
-
-class Godfather(Mafia):
-    def __init__(self, nickname):
-        super(Godfather, self).__init__(nickname)
-
-    def sheriffs_check(self, nickname):
-        pass
-
-
-if __name__ == '__main__':
-    players = [Mafia(""), Godfather(""), Citizen(""), Citizen(""), Citizen(""), Mafia(""), Sheriff("")]
-    for player in players:
-        if player is Mafia:
-            player.kill_player("")
-
